refactor: move enhancement to enhancements directory
